requirement.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
colorlog==5.0.1
har2case==0.3.1
HttpRunner==1.5.8
idna==2.10
iniconfig==1.1.1
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
pytest-assume==2.4.3
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.4
